requirements.txt
alembic==1.9.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Migrate==4.0.1
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
python-decouple==3.7
requests==2.28.2
SQLAlchemy==1.4.46
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.11.0
